move regression.tests results files to regression.tests.results
