+++ statsmodels/regression/tests/results/generate_lasso.py
