+++ statsmodels/regression/tests/results/glmnet_r_results.py
