+++ statsmodels/regression/tests/results/lasso_r_results.R
