+++ statsmodels/regression/tests/results/lme_r_results.py
